Initialising the rest of the files
Firstly established the rest of the model files. Modified some of the relationships between Trail and the rest of the different tables as it wasn't clear enough.
Now trying to implement the routes into the main app.py, and create a database using a python script with some examples of data